Update raw.rs
Apparently on the Raspberry pi (I think it's a 3B that I have) std::os::raw::c_char which is used by bindgen when generating the ffi functions is a u8. 
Because of this the buf.as_mut_ptr() as *mut i8 in the get_string function wouldn't compile on the raspberry pi. This simple change allows it to compile.